:recycle: Combine subprocess test files
